requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anthropic==0.3.6
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
boilerpy3==1.0.6
Brotli==1.0.9
cachetools==5.3.1
canals==0.3.2
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.6
coloredlogs==15.0.1
comm==0.1.3
construct==2.5.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.13
debugpy==1.6.7
decorator==5.1.1
distro==1.8.0
docopt==0.6.2
duckduckgo-search==3.8.4
Events==0.4
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
farm-haystack==1.19.0
fastapi==0.99.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-search-results==2.4.2
h11==0.14.0
h2==4.1.0
haystack==0.42
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface==0.0.1
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflect==7.0.0
install==1.3.5
InstructorEmbedding==1.0.1
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
langchain==0.0.242
langsmith==0.0.14
lazy-imports==0.3.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==10.0.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdfminer==20191125
pdfminer.six==20221105
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
plotly==5.15.0
posthog==3.0.1
prompt-toolkit==3.0.39
prompthub-py==4.0.0
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.13.0
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ptrace==0.9.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-cache==0.9.8
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sseclient-py==1.7.2
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
url-normalize==1.4.3
urllib3==2.0.4
uvicorn==0.23.1

uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2

anthropic